+++ spec/app/models/location.rb
@@ -2,5 +2,6 @@ class Location
   include Mongoid::Document
   field :name
   field :info, :type => Hash
+  field :occupants, :type => Array
   embedded_in :address
 end