+++ lib/mongoid/fields/internal/array.rb
@@ -46,29 +46,8 @@ module Mongoid #:nodoc:
         #
         # @since 2.1.0
         def serialize(object)
-          raise_or_return(object)
-        end
-
-        protected
-
-        # If the value is not an array or nil we will raise an error,
-        # otherwise return the value.
-        #
-        # @example Raise or return the value.
-        #   field.raise_or_return([])
-        #
-        # @param [ Object ] value The value to check.a
-        #
-        # @raise [ InvalidType ] If not passed an array.
-        #
-        # @return [ Array ] The array.
-        #
-        # @since 2.1.0
-        def raise_or_return(value)
-          unless value.nil? || value.is_a?(::Array)
-            raise Mongoid::Errors::InvalidType.new(::Array, value)
-          end
-          value
+          return object unless object
+          ::Array.wrap(object)
         end
       end
     end